fix: mismatched scraped item

+++ scrape.py
@@ -66,49 +66,65 @@ def scrape(universe):
 
     start_time = time.time()
 
-    response = requests.post(
-        "https://apac1.apps.cp.thomsonreuters.com/apps/udf/msf/", data=json.dumps(request_payload), headers=headers)
+    if os.path.exists("json/{}.json".format(universe)):
+        with open("json/{}.json".format(universe), "r", encoding="utf-8") as fp:
+            data = json.load(fp)
+    else:
+        response = requests.post(
+            "https://apac1.apps.cp.thomsonreuters.com/apps/udf/msf/", data=json.dumps(request_payload), headers=headers)
 
-    if response.status_code != 200:
-        print("ERROR: ", response.content)
+        if response.status_code != 200:
+            print("ERROR: ", response.content)
 
-    try:
-        data = response.json()
-    except Exception:
-        print("ERROR: unable to fetch data! Please login and grab the latest cookie!")
-        exit(1)
+        try:
+            data = response.json()
+        except Exception:
+            print("ERROR: unable to fetch data! Please login and grab the latest cookie!")
+            exit(1)
 
-    with open("json/{}.json".format(universe), "w", encoding="utf-8") as fp:
-        json.dump(data, fp)
+        with open("json/{}.json".format(universe), "w", encoding="utf-8") as fp:
+            json.dump(data, fp)
 
     results = data['r']
 
-    idx = 0
-    for formula in tqdm(formulas[:-2], desc="Processing formulas..."):
-        topic_name = formula.split(".")[-1]
-        topic_path = os.path.join("data", topic_name)
-        if not os.path.exists(topic_path):
-            os.mkdir(topic_path)
-        start = idx * 7 + 1
-        end = start + 7
-        for i in range(start, end):
-            if len(results) <= i:
-                continue
+    if len(results) <= 1:
+        print("NO DATA FOR {}".format(universe))
+        print(data)
+        pass
 
-            abstract = results[i]["v"]
-            year = results[i]["v"][5].get("___MSFVALUE", 0)
-            if len(abstract) <= 17:
+    else:
+        prev_year = None
+        offset = 0  # account for duplicated year data
+        for idx in tqdm(range(1, len(results)), desc="Processing formulas..."):
+            v = results[idx]["v"]
+            year = v[5].get("___MSFVALUE", 0)
+            if year == 0:
                 continue
-            abstract = abstract[17].get("___MSFVALUE", None)
-            if abstract is None:
+            if prev_year is not None and year == prev_year:
+                offset += 1
+                continue
+            prev_year = year
+            if len(v) <= 18:
+                continue
+            abstract = v[17].get("___MSFVALUE", None)
+            if abstract is None or abstract == "English":
+                abstract = v[18].get("___MSFVALUE", None)
+            if abstract is None or abstract == "False":
                 continue
             if len(abstract) < 2:
                 continue
-            abstract = abstract.replace("\n", " ")
+            abstract = abstract.replace(
+                "\n", " ")
 
+            topic_idx = (idx - offset - 1) // 7
+            topic_name = formulas[:-2][topic_idx].split(".")[-1]
+            topic_path = os.path.join(
+                "data", topic_name)
+            if not os.path.exists(topic_path):
+                os.mkdir(topic_path)
+            print(topic_name, idx, "offset:", offset)
             with open(topic_path+"/{}-{}.txt".format(year, universe), "w", encoding="utf-8") as fh:
                 fh.write(abstract)
-        idx += 1
 
     print("Done scraping for {}".format(universe))
